pyproject.toml
[project]
name = "smartpick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "langchain-core>=1.2.11",
    "langchain-upstage>=0.7.6",
    "langgraph>=1.0.2",
    "openai>=2.2.0",
    "pandas>=2.2.0",
    "pdfplumber>=0.11.9",
    "playwright>=1.58.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.35.0",
]
